Move eval tests to eval/tests.rs.
